+++ src/components/league/index.tsx
@@ -7,7 +7,9 @@ import Spinner from '../spinner';
 import mandali from '../mandali.module.css';
 
 const SplitedValues = Dynamic(() => import('./boxes/SplitedExalted'), { loading: () => <Spinner/> });
-const ChangeHelper = Dynamic(() => import('./boxes/ChangeHelper'), { loading: () => <Spinner/> });
+const ChangeHelper = Dynamic(() => import('./boxes/change-helper'), {
+  loading: () => <Spinner />,
+});
 const TableView = Dynamic(() => import('./table'), { loading: () => <Spinner/> });
 
 export default function League() {
